FetchBindex/requirements.txt
Pillow==2.9.0
selenium==2.53.6
requests==2.2.1
openpyxl==2.3.3
xlwt==1.0.0
chardet==3.0.2
threadpool==1.3.2